mock_api/pyproject.toml
[project]
name = "mock-api"
version = "0.1.0"
description = "Mock API server for database operations"
requires-python = ">=3.9"
dependencies = [
    "flask>=2.3.0",
    "flask-cors>=4.0.0",
    "psycopg2-binary>=2.9.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]

[project.scripts]
mock-api = "mock_api.server:main"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]
